sample: Organize code related to RecyclerView tests
