+++ sample/build.gradle
@@ -31,11 +31,16 @@ android {
 
 dependencies {
     implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk7:$versions.kotlin"
+
     implementation "com.android.support:appcompat-v7:$versions.supportLibrary"
     implementation "com.android.support:support-v4:$versions.supportLibrary"
     implementation "com.android.support:recyclerview-v7:$versions.supportLibrary"
     implementation "com.android.support:design:$versions.supportLibrary"
+    implementation "com.android.support:support-annotations:$versions.supportLibrary"
     implementation "com.android.support.constraint:constraint-layout:$versions.constraintLayout"
+
+    implementation "com.google.code.findbugs:jsr305:$versions.findBugsAnnotations"
+
     testImplementation "junit:junit:$versions.jUnit"
     androidTestImplementation project(':kappuccino')
     androidTestImplementation "com.android.support.test.espresso:espresso-intents:$versions.espresso"
