+++ sample/src/main/AndroidManifest.xml
@@ -25,6 +25,7 @@
         <activity android:name=".RuntimePermissionActivity" />
         <activity android:name=".CompoundDrawablesActivity"/>
         <activity android:name=".MenuActivity"/>
+        <activity android:name=".recyclerview.RecyclerViewActivity"/>
 
     </application>
