+++ sample/src/main/java/br/com/concretesolutions/kappuccino/sample/recyclerview/RecyclerViewAdapter.java
@@ -1,0 +1,58 @@
+package br.com.concretesolutions.kappuccino.sample.recyclerview;
+
+import android.support.annotation.NonNull;
+import android.support.v7.widget.RecyclerView;
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+import android.widget.TextView;
+
+import java.util.ArrayList;
+
+import javax.annotation.ParametersAreNonnullByDefault;
+
+import br.com.concretesolutions.kappuccino.sample.R;
+
+@ParametersAreNonnullByDefault
+public class RecyclerViewAdapter extends RecyclerView.Adapter<RecyclerViewAdapter.ViewHolder>{
+
+    private final ArrayList<String> subjectValues;
+
+    RecyclerViewAdapter(ArrayList<String> subjectValues) {
+        this.subjectValues = subjectValues;
+    }
+
+    public void removeAt(int position) {
+        subjectValues.remove(position);
+        notifyItemRemoved(position);
+    }
+
+    @Override
+    public @NonNull RecyclerViewAdapter.ViewHolder onCreateViewHolder(ViewGroup parent, int viewType) {
+        View view = LayoutInflater
+                .from(parent.getContext())
+                .inflate(R.layout.recyclerview_items, parent, false);
+        return new ViewHolder(view);
+    }
+
+    @Override
+    public void onBindViewHolder(ViewHolder holder, int position) {
+        holder.textView.setText(subjectValues.get(position));
+    }
+
+    @Override
+    public int getItemCount(){
+        return subjectValues.size();
+    }
+
+    @ParametersAreNonnullByDefault
+    static class ViewHolder extends RecyclerView.ViewHolder {
+
+        private TextView textView;
+
+        ViewHolder(View v) {
+            super(v);
+            textView = v.findViewById(R.id.subject_textview);
+        }
+    }
+}
